third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_innerHTML_adoption01-expected.txt

This is a testharness.js-based test.
[FAIL] html5lib_innerHTML_adoption01.html 0bf80e1546d4c221354aa9734f61713b7d64ee6d
  assert_equals: expected "#document\\n| <b>\\n|   <em>\\n|     <foo>\\n|       <foob>\\n|         <fooc>\\n| <aside>\\n|   <b>" but got "#document\\n| <b>\\n|   <em>\\n|     <foo>\\n|       <foob>\\n|         <fooc>\\n| <aside>\\n|   <em>\\n|     <b>"
Harness: the test ran to completion.
